wmaker patches

+++ WPrefs.app/MouseSettings.c
@@ -298,10 +298,10 @@ static void showData(_Panel * panel)
 	sprintf(buffer, "%i", a);
 	WMSetTextFieldText(panel->threT, buffer);
 
-	WMSetSliderValue(panel->speedS, (accel - 0.25) / 0.25);
+	WMSetSliderValue(panel->speedS, (accel - 0.25F) / 0.25F);
 
 	panel->acceleration = accel;
-	sprintf(buffer, "%.2f", accel);
+	sprintf(buffer, "%.2f", (double)accel);
 	WMSetTextFieldText(panel->acceT, buffer);
 
 	 /**/ b = GetIntegerForKey("DoubleClickTime");
